Refactor by creating `XLabelPainter`
